Credit fool

See merge request enigma-camp/malang-branch/torvalds-batch-4/team/final-projects/team-4/giri-guide-backend!53

+++ src/main/java/com/abdav/giri_guide/service/impl/CustomerServiceImpl.java
@@ -88,7 +88,9 @@ public class CustomerServiceImpl implements CustomerService {
         ImageEntity img = imageService.create(file, PathImage.PROFILE_PICTURE, customer.getFullName());
         customer.setImage(img);
         customerRepository.saveAndFlush(customer);
-        imageService.delete(oldImage);
+        if (oldImage != null) {
+            imageService.delete(oldImage);
+        }
 
         return CustomerMapper.customerToCustomerResponse(customer, httpReq);
     }
